Rehouse bounce height tests
